feat: consider timeline date change during selection
