+++ frontend/src/features/map_planning/store/UntrackedMapStore.ts
@@ -6,6 +6,7 @@ import {
   UNTRACKED_DEFAULT_STATE,
   UntrackedMapSlice,
 } from './MapStoreTypes';
+import { handleSelectedNodesChange } from './utils';
 import { FrontendOnlyLayerType } from '@/features/map_planning/layers/_frontend_only';
 import Konva from 'konva';
 import { Vector2d } from 'konva/lib/types';
@@ -180,6 +181,8 @@ export const createUntrackedMapSlice: StateCreator<
           fetchDate: date,
         },
       }));
+
+      handleSelectedNodesChange(get);
       return;
     }
 
@@ -201,6 +204,7 @@ export const createUntrackedMapSlice: StateCreator<
         },
       },
     }));
+    handleSelectedNodesChange(get);
   },
   /**
    * Allow the user to make measurement inputs on the base layer.
